split Modal
